Refactoring
- Brings api requests in separate file
- Rename data reducer
- Rename reducer folder
- Rename handlers in cities-places
- Check default city in cities places after mount not after update
