+++ src/api/clients.js
@@ -1,0 +1,2 @@
+export const getHotels = (api) => api.get(`/hotels`);
+export const getAuthorizationStatus = (api) => api.get(`/login`);
