+++ src/components/cities-list/cities-list.jsx
@@ -1,36 +1,48 @@
-import React from "react";
+import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
 import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer/page/page.js";
-import {getActiveCity} from "../../reducer/page/selectors.js";
-import {getCities} from "../../reducer/data/selectors.js";
+import {ActionCreator} from "../../redux/page/page.js";
+import {getActiveCity} from "../../redux/page/selectors.js";
+import {getCities} from "../../redux/offersData/selectors.js";
 
 
-function CitiesList({city: activeCity, onChooseCityClick, cities, setDefaultCity}) {
-  return (
-    <div className="tabs">
-      <section className="locations container">
-        <ul className="locations__list tabs__list">
-          {activeCity === null && setDefaultCity(cities[0])}
-          {cities.map((city) => {
-            const onClick = () => onChooseCityClick(city);
-            return (
-              <li
-                className="locations__item"
-                key={city}
-                onClick={onClick}>
-                <a
-                  className={`locations__item-link tabs__item  ${city === activeCity ? `tabs__item--active` : ``}`}
-                  href="#">
-                  <span>{city}</span>
-                </a>
-              </li>
-            );
-          })}
-        </ul>
-      </section>
-    </div>
-  );
+class CitiesList extends PureComponent {
+  constructor(props) {
+    super(props);
+
+  }
+
+  componentDidMount() {
+    const {city: activeCity, cities, setDefaultCity} = this.props;
+    return activeCity === null && setDefaultCity(cities[0]);
+  }
+
+  render() {
+    const {city: activeCity, onChooseCityClick, cities} = this.props;
+    return (
+      <div className="tabs">
+        <section className="locations container">
+          <ul className="locations__list tabs__list">
+            {cities.map((city) => {
+              const handleChooseCityClick = () => onChooseCityClick(city);
+              return (
+                <li
+                  className="locations__item"
+                  key={city}
+                  onClick={handleChooseCityClick}>
+                  <a
+                    className={`locations__item-link tabs__item  ${city === activeCity ? `tabs__item--active` : ``}`}
+                    href="#">
+                    <span>{city}</span>
+                  </a>
+                </li>
+              );
+            })}
+          </ul>
+        </section>
+      </div>
+    );
+  }
 }
 
 CitiesList.propTypes = {
