+++ src/index.js
@@ -7,9 +7,9 @@ import thunk from "redux-thunk";
 import {Provider} from "react-redux";
 import {createAPI} from "./api/api.js";
 import App from "./components/app/app.jsx";
-import {Operation as DataOperation} from "./reducer/data/data.js";
-import {ActionCreator, AuthorizationStatus} from "./reducer/user/user.js";
-import reducer from "./reducer/reducer.js";
+import {Operation as DataOperation} from "./redux/offersData/offersData.js";
+import {ActionCreator, AuthorizationStatus} from "./redux/user/user.js";
+import reducer from "./redux/reducers.js";
 
 const onUnauthorized = () => {
   store.dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH));
