+++ src/redux/offersData/selectors.js
